Move "Report this add-on" to "More information" section, make it a link
