+++ src/amo/components/AddonReportAbuseLink/index.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import { compose } from 'redux';
 
 import translate from 'amo/i18n/translate';
-import Button from 'amo/components/Button';
+import Link from 'amo/components/Link';
 import type { AddonAbuseState } from 'amo/reducers/abuse';
 import type { AppState } from 'amo/store';
 import type { AddonType } from 'amo/types/addons';
@@ -31,7 +31,7 @@ type InternalProps = {|
   i18n: I18nType,
 |};
 
-export class ReportAbuseButtonBase extends React.Component<InternalProps> {
+export class AddonReportAbuseLinkBase extends React.Component<InternalProps> {
   render(): null | React.Node {
     const { abuseReport, addon, i18n, loading } = this.props;
 
@@ -39,12 +39,12 @@ export class ReportAbuseButtonBase extends React.Component<InternalProps> {
 
     if (abuseReport && abuseReport.message !== undefined) {
       return (
-        <div className="ReportAbuseButton ReportAbuseButton--report-sent">
-          <h3 className="ReportAbuseButton-header">
+        <div className="AddonReportAbuseLink AddonReportAbuseLink--report-sent">
+          <h3 className="AddonReportAbuseLink-header">
             {i18n.gettext('You reported this add-on')}
           </h3>
 
-          <p className="ReportAbuseButton-first-paragraph">
+          <p className="AddonReportAbuseLink-first-paragraph">
             {i18n.gettext(`We have received your report. Thanks for letting us
               know about your concerns with this add-on.`)}
           </p>
@@ -52,24 +52,13 @@ export class ReportAbuseButtonBase extends React.Component<InternalProps> {
       );
     }
 
-    const reportButtonProps: Object = {
-      to: `/feedback/addon/${addon.slug}/`,
-    };
-
     /* eslint-disable react/no-danger */
     return (
-      <div className={makeClassName('ReportAbuseButton')}>
-        <div className="ReportAbuseButton--preview">
-          <Button
-            buttonType="neutral"
-            className="ReportAbuseButton-show-more"
-            disabled={loading}
-            rel="nofollow"
-            puffy
-            {...reportButtonProps}
-          >
+      <div className={makeClassName('AddonReportAbuseLink')}>
+        <div className="AddonReportAbuseLink--preview">
+          <Link to={`/feedback/addon/${addon.slug}/`} rel="nofollow">
             {i18n.gettext('Report this add-on')}
-          </Button>
+          </Link>
         </div>
       </div>
     );
@@ -89,9 +78,9 @@ const mapStateToProps = (state: AppState, ownProps: Props): PropsFromState => {
   };
 };
 
-const ReportAbuseButton: React.ComponentType<Props> = compose(
+const AddonReportAbuseLink: React.ComponentType<Props> = compose(
   connect(mapStateToProps),
   translate(),
-)(ReportAbuseButtonBase);
+)(AddonReportAbuseLinkBase);
 
-export default ReportAbuseButton;
+export default AddonReportAbuseLink;
