+++ src/amo/components/AddonReportAbuseLink/styles.scss
@@ -1,0 +1,18 @@
+@import '~amo/css/styles';
+
+.AddonReportAbuseLink {
+  margin: 24px auto 0;
+}
+
+.AddonReportAbuseLink-header {
+  margin-bottom: 0;
+}
+
+.AddonReportAbuseLink-first-paragraph {
+  margin-top: 6px;
+}
+
+.AddonReportAbuseLink--preview {
+  display: inline-block;
+  width: 100%;
+}
